refactor: move modules path
