+++ src/discord/modules/help/help.json
